==> 01_real_estate_preprocessing.ipynb <==
# Part 01
----
아파트 매매 데이터에서 불필요한 칼럼들을 제거한 뒤 추후 인구이동 데이터와 비교분석을 위해 주소 간략화 및 행자부코드를 부여한다.

행자부코드 목록에 명기된 주소보다 아파트 매매 데이터에는 더 상세히 기술되어 있어 주소 문자열을 처음 두 단어 기준으로 간략화 시켰다. 따라서 일반적으로 대도시인 특별 행정구역의 경우 시+구의 단위로 쓰이게 되며 그 외에는 도+시 단위로 쓰이게 된다.

세종특별시의 경우 생긴지 얼마 안된 이유로 잦은 행정구역의 변화가 있었기에 임의로 이전 행정구역을 사용하도록 수정을 한다.

마지막으로 건축년도가 null값으로 되어있는 데이터 역시 수정을 가해 null값이 없는 깨끗한 데이터셋을 만든다.

%matplotlib inline
import re

# set path and read files
path = 'raw_data\\'
df_estate = pd.read_csv(path + 'real_estate_2006_2015.csv', encoding='utf-8')
df_address_code = pd.read_excel(path + 'address_code.xlsx', header=2)

# have a quick glance
df_estate.tail()

# rename columns into english
estate_column_dic = {
    '시군구': 'address',
    '번지': 'block',
    '본번': 'block_main',
    '부번': 'block_sub',
    '단지명': 'apt',
    '전용면적(㎡)': 'area',
    '계약년월': 'year_month',
    '계약일': 'day',
    '거래금액(만원)': 'price',
    '층': 'floor',
    '건축년도': 'built',
    '도로명': 'street'
}

# rename column names into English
df_estate.rename(columns=estate_column_dic, inplace=True)

# list of columns that we will keep
df_columns = ['address', 'area', 'year_month', 'price', 'built']

# drop all columns we don't need
df_estate = df_estate[df_columns]

df_estate.head()

# convert prices into integer type
df_estate.price = df_estate.price.str.replace(',', '').astype(int)

# we split up year and month sold to separate columns
df_estate['year'] = df_estate.year_month // 100
df_estate['month'] = df_estate.year_month % 100
df_estate.drop('year_month', axis=1, inplace=True)

# as year built is of float type, we perform a sanity check
# fortunately, there are only four with missing values
df_estate[df_estate.built.isnull()]

# one apartment can be reasonably assumed to be a part of the apartment complex called 태하
df_estate.loc[df_estate[df_estate.built.isnull()].index[:-1], 'built'] = 1987

# we also drop a single row which cannot be identified
df_estate = df_estate.loc[~df_estate.built.isnull()]

# there is one apartment with built year set to zero
df_estate[df_estate.built == 0].head()

# after some investigation, we set this to 1976 and convert the column into integer type
df_estate.loc[df_estate.built == 0, 'built'] = 1976
df_estate.built = df_estate.built.astype(int)

# check the changes made
df_estate.head()

# from here, we try to convert address texts to area codes

# function that returns the first word or two words of the string
subaddress_regex = re.compile(u'(\S+\s|\S+)\s*(\S*)')
def shrink_address(address):
    return ''.join(subaddress_regex.search(address).groups())

# remove whitespaces at the beginning of strings and at the end of strings,
# shrink multiple white spaces, and cut off addresses after the first two words
df_estate.address = df_estate.address.apply(shrink_address)

df_estate.head()

# similar process as above
df_address_code['address'] = df_address_code['행정구역명'].apply(shrink_address)

df_address_code['area_code'] = df_address_code['행자부코드'] // 100000

# as we have both shortened address texts and area codes,
# we get duplicates in both columns
# when creating maps, we only take the first or the last among the duplicates
# the reason for this is to update old area codes, which appears last, to new ones
# through the following process:
# old area code -> address -> new area code

df_address_to_code = df_address_code[['address', 'area_code']] \
    .drop_duplicates(subset='address', keep='first').set_index('address')

df_code_to_address = df_address_code[['address', 'area_code']] \
    .drop_duplicates(subset='area_code', keep='last').set_index('area_code')

# now create a column named area code
df_estate['area_code'] = df_estate.address.map(df_address_to_code.area_code)

# we see below address codes are given float values instead of integer values
# this probably means there are some addresses that did not get mapped properly
# hence getting assigned nan (float) values
df_estate.tail()

# we check to see which addresses are not taken care of properly
# fortunately, there are only three
df_estate[df_estate.area_code.isnull()].groupby('address').count()

# getting ready to patch things up manually
# 고운동, 나성동, 어진동 were all part of 한솔동
# moreover, 도담동, 아름동, 종촌동 were all part of 한솔동 until very recently
# so we might as well combine everything into one address
df_estate.address.replace(to_replace=['세종특별자치시 고운동', '세종특별자치시 나성동', '세종특별자치시 어진동',
                                      '세종특별자치시 도담동', '세종특별자치시 아름동', '세종특별자치시 종촌동'],
                          value='세종특별자치시 한솔동', inplace=True)

# try mapping string addresses to codes again
df_estate['area_code'] = df_estate.address.map(df_address_to_code.area_code)

df_estate.tail()

# one last sanity check
df_estate.isnull().values.any()

# we don't need the address column anymore
df_estate.drop('address', axis=1, inplace=True)

# export to a new file
df_estate.to_csv(path + '01_estate.csv', index=False, encoding='utf-8')

# export the conversion tables for later use
df_address_to_code.to_csv(path + '01_address_to_code.csv', index=True, encoding='utf-8')
df_code_to_address.to_csv(path + '01_code_to_address.csv', index=True, encoding='utf-8')
