==> 02_population_preprocessing.ipynb <==
# Part 02
----
아래에서는 인구이동 데이터의 1차적인 전처리 과정을 행한다. 우선적으로 읍면동 부분을 제외한 행자부코드를 이용하여 하나의 지역코드로 묶는다. 이후 불필요한 칼럼을 없앤다.

2012년과 2013년의 데이터의 경우 이동 총인구가 마침표로 잘못 기입된 데이터가 있어 이를 수정하고 정수타입으로 변환한다.

또한 누락된 데이터 외의 모든 데이터는 정수타입으로 표기될 수 있으므로 nan값을 -1로 수정 한뒤 정수타입으로 변환시킨다.

세대주 성별의 경우 주민등록상 등록된 번호로 구분이 되어 있기에 일괄적으로 대응되는 1 또는 2로 변환한다.

세대주 구분에서 0과 -1은 모두 비세대주로 간주하여 0으로 통일한다.

마지막으로 데이터 탐색 단계에서 이동 총인구가 -1이거나나 0으로 기입된 데이터 포인트가 119개 있다는 것을 알수 있었으며, 이 경우 이동 남인구와 이동 여인구의 데이터가 모두 0으로 되어있는 것을 감안, 그리고 전체 데이터 크기인 약 6300만과 비교하여 상당히 적은 수인 것도 감안하여 이동 총인구가 0보다 작거나 같은 경우 누락시키기로 한다.

위의 전처리 과정을 위한 데이터 탐구 과정(이동 총인구 데이터의 타입, 세대주 성별 구분 코드, 세대주 관계 코드의 비일관성 발견 등)은 아래 코드에는 생략한다.

%matplotlib inline
import glob

path = 'raw_data\\'

population_columns = ['in_state', 'in_city', 'in_nbr', 'year', 'month', 'in_date',
                      'out_state', 'out_city', 'out_nbr', 'reason',
                      'hh_head', 'head_age', 'head_male', 'mov_multi',
                      'mov_ppl', 'mov_male', 'mov_female']

# we drop move the following columns:
#     date: won't be used as it is too specific
#     nbr: same reason as above
#     mov_multi, mov_ppl: can be derived from mov_male and mov_female
used_columns = ['in_state', 'in_city', 'year', 'month',
                'out_state', 'out_city', 'reason',
                'hh_head', 'head_age', 'head_male',
                'mov_male', 'mov_female']

population_files = glob.glob(path + 'population_*.txt')

# take a random file to fiddle around with
df = pd.read_csv(population_files[np.random.randint(len(population_files))],
                 names=population_columns, usecols=used_columns)

## Data Scrubbing

df.head()

# checking null values
df.isnull().any()

# replace all null values to -1
df = df.replace(to_replace=np.NaN, value=-1)

# we convert everything into int
df = df.astype(int)

# create a column with combined address codes
df['in_address'] = df.in_state * 1000 + df.in_city
df['out_address'] = df.out_state * 1000 + df.out_city
df = df.drop(['in_state', 'in_city', 'out_state', 'out_city'], axis=1)

df.head()

# anomalies below were discovered much later in the process
# and are spread out through different files
# we leave these here for the sake of completeness

# change gender codes to 1 (male) or 0 (female) accordingly
# this is due to how gender specifying code works in korean id system
df.loc[(df.head_male == 9) | (df.head_male == 3), 'head_male'] = 1
df.loc[df.head_male != 1, 'head_male'] = 0

# change the missing data values -1 (originally null) to 0 for consistency
# there are cases where values were simply missing rather than being marked as 0
df.loc[df.hh_head == -1, 'hh_head'] = 0

# final check
df.sample(5)

## Data Aggregation

# make a new column indicating whether the movement happend within the are or not
df['foreign'] = (df.in_address != df.out_address) * 1

# we derive these two columns from others
df['mov_ppl'] = df.mov_male + df.mov_female
df['mov_multi'] = (df.mov_ppl > 1) * 1
df['house_total'] = 1

# there are rare occasions where the data of number of people moved is missing
# there are exactly 116 entries of this kind, and as we have more than 63,000,000
# data points, we simply remove them
df = df.loc[df.mov_ppl > 0]

df.sample(5)

# defining a custom function to use when aggregating dataframes
# when calculating the average of the ages, we only count those greater than or equal to 0
# as missing values are indicated by -1
def age_average(series):
    if series[series >= 0].count() != 0:
        return series[series >= 0].sum() / series[series >= 0].count()
    else:
        return -1

# this is a dictionary that will be used to aggregate the original data
agg_functions = {
    'hh_head': 'sum',
    'head_age': age_average,
    'head_male': 'sum',
    'mov_multi': 'sum',
    'mov_ppl': 'sum',
    'mov_male': 'sum',
    'house_total': 'count',
}

# we create two dataframes, one for people moving into area codes and
# another for people moving out of area codes
# this way, although each data point gets 'duplicated', we do this so that we may
# compute population changes for area codes (and other stuff) just in case if they become needed
df_i = df.groupby(['year', 'month', 'in_address', 'foreign', 'reason'], as_index=False).agg(agg_functions)
df_o = df.groupby(['year', 'month', 'out_address', 'foreign', 'reason'], as_index=False).agg(agg_functions)

# add in a new column to indicate whether the data is for people moving in or out
df_i['move_in'] = 1
df_o['move_in'] = 0

df_i.sample(5)

df_o.sample(5)

io_rename_dic = {'in_address': 'area_code', 'out_address': 'area_code'}

# rename the columns for consistency to support concatenation later
df_i = df_i.rename(columns=io_rename_dic)
df_o = df_o.rename(columns=io_rename_dic)

# note that mov_female column is dropped since we can derive it from other columns
io_column_order = ['year', 'month', 'area_code', 'move_in', 'foreign', 'reason',
                   'house_total', 'hh_head', 'head_age', 'head_male', 'mov_multi',
                   'mov_ppl', 'mov_male']

# make sure columns are in the same order just in case
df_i = df_i[io_column_order]
df_o = df_o[io_column_order]
df = pd.concat([df_i, df_o])

# check the final result
df.sample(5)

## Automation

# in the following section below, we combine the codes above
# and turn them into functions so that we may automate the whole process

# takes a population migration data and apply scrubbing
def pop_scrub(df):
    df = df.replace(to_replace=np.NaN, value=-1)
    df = df.astype(int)
    
    df['in_address'] = df.in_state * 1000 + df.in_city
    df['out_address'] = df.out_state * 1000 + df.out_city
    df = df.drop(['in_state', 'in_city', 'out_state', 'out_city'], axis=1)
    
    df.loc[(df['head_male'] == 9) | (df['head_male'] == 3), 'head_male'] = 1
    df.loc[df.head_male != 1, 'head_male'] = 0
    df.loc[df.hh_head == -1, 'hh_head'] = 0
    
    return df

# groups the migration data by year, month, area code, migrate, and reason
def pop_agg(df):
    df['foreign'] = (df.in_address != df.out_address) * 1
    df['mov_ppl'] = df.mov_male + df.mov_female
    df['mov_multi'] = (df.mov_ppl > 1) * 1
    df['house_total'] = 1
    df = df.loc[df.mov_ppl > 0]
    
    df_i = df.groupby(['year', 'month', 'in_address', 'foreign', 'reason'], as_index=False).agg(agg_functions)
    df_o = df.groupby(['year', 'month', 'out_address', 'foreign', 'reason'], as_index=False).agg(agg_functions)
    
    df_i['move_in'] = 1
    df_o['move_in'] = 0
    df_i = df_i.rename(columns=io_rename_dic)
    df_o = df_o.rename(columns=io_rename_dic)
    df_i = df_i[io_column_order]
    df_o = df_o[io_column_order]
    
    return pd.concat([df_i, df_o])

df = []
for population_file in population_files:
    df_temp = pd.read_csv(population_file, names=population_columns, usecols=used_columns)
    
    df_temp = pop_scrub(df_temp)
    df_temp = pop_agg(df_temp)
    
    df.append(df_temp)
    # print statements to check progress
    print(population_file[-8:-4], end=' ')

pd.concat(df).to_csv(path + '02_pop.csv', index=False)
