==> 03_fix_area_codes.ipynb <==
# Part 03
----
아파트 매매와 인구이동 데이터의 비교분석을 용이하게 하기 위한 전처리 과정을 행한다.

우선적으로 년월의 구분 방법이 다른 것을 같은 형태로 만든다. 또한 데이터 조작의 편의를 위해서 칼럼 이름을 영문으로 전부 고치고 비교 가능한 칼럼들은 같은 이름으로 통일한다.

이후 대부분의 코드는 행정코드의 불일치를 고치는 과정이다. 행정코드의 불일치는 크게 두가지 이유로 남아있다.

하나는 아파트 매매의 경우 주소를 무조건적으로 처음 두단어로 축약하여 코드로 매핑을 하였지만 인구 이동의 경우 행정코드 앞 다섯자리만 남기는 방식으로 전처리를 거쳤기 때문에 방식의 차이로 인해 코드 불일치가 일어났다. 일반적으로 인구 이동 데이터에서의 행정코드가 더 세밀한 단위까지 구분되어 있다. 이를 해결하기 위해 인구 이동 데이터에서 행정코드 -> 주소 -> 축약주소 -> 행정코드 과정을 거쳐 최대한 아파트 매매의 행정코드와 일치시킨다.

나머지 하나는 행정구역의 변화로 인하여 두개의 데이터 프레임은 같은 지역임에도 서로 다른 행정코드를 가지고 있는 경우도 있으며 다행히 이 경우는 약 10가지 경우 남짓하여 손으로 직접 수정을 가한다.

%matplotlib inline
import re

# setting the folder path and read in data
path = 'raw_data\\'
df_pop = pd.read_csv(path + '02_pop.csv', encoding='utf-8')
df_estate = pd.read_csv(path + '01_estate.csv', encoding='utf-8')

# quick initial check
df_pop.head()

df_estate.head()

estate_column_order = [
    'year', 'month', 'area_code', 'area', 'price', 'built'
]

df_estate = df_estate[estate_column_order]

df_estate.head()

# we see that the number of area codes do not match up between the dataframes
df_pop.groupby('area_code').count().shape

df_estate.groupby('area_code').count().shape

# we try to see the difference in the area codes used in both dataframes
estate_area_code = df_estate.area_code[~df_estate.area_code.duplicated()]
estate_area_code_missing = estate_area_code[~estate_area_code.isin(df_pop.area_code)]
estate_area_code_missing.values

pop_area_code = df_pop.area_code[~df_pop.area_code.duplicated()]
pop_area_code_missing = pop_area_code[~pop_area_code.isin(df_estate.area_code)]
pop_area_code_missing.values

# just a custom function to use as an aggregate method
def agg_nunique(series):
    return series.nunique()

address_to_code = pd.read_csv(path + '01_address_to_code.csv', index_col=0).area_code
code_to_address = pd.read_csv(path + '01_code_to_address.csv', index_col=0).address

# function that returns the first word or two words of the string
subaddress_regex = re.compile(u'(\S+\s|\S+)\s*(\S*)')
def shrink_address(address):
    return ''.join(subaddress_regex.search(address).groups())

pd.DataFrame(data={
    'area_code': estate_area_code_missing.values,
    'address': estate_area_code_missing.map(code_to_address).apply(shrink_address)
})

# comparing with the table above, we see that in many cases, the problem arises from
# the area codes being too granular in the population dataframe
pd.DataFrame(data={
    'area_code': pop_area_code_missing.values,
    'address': pop_area_code_missing.map(code_to_address).apply(shrink_address)
})

# run the address code through the process explained above
temp_address = df_pop.area_code.map(code_to_address)
temp_address = temp_address.apply(shrink_address)
temp_address = temp_address.map(address_to_code)
df_pop.area_code = temp_address

# we repeat the process as before trying to see the changes
estate_area_code = df_estate.area_code[~df_estate.area_code.duplicated()]
estate_area_code_missing = estate_area_code[~estate_area_code.isin(df_pop.area_code)]
estate_area_code_missing.values

pd.DataFrame(data={
    'area_code': estate_area_code_missing.values,
    'address': estate_area_code_missing.map(code_to_address).apply(shrink_address)
})

pop_area_code = df_pop.area_code[~df_pop.area_code.duplicated()]
pop_area_code_missing = pop_area_code[~pop_area_code.isin(df_estate.area_code)]
pop_area_code_missing.values

pd.DataFrame(data={
    'area_code': pop_area_code_missing.values,
    'address': pop_area_code_missing.map(code_to_address).apply(shrink_address)
})

# as there are only 11 values left to fix, we tidy up things manually
# 옹진군 is actually a collection of small islands, hence it is reasonable there are no appartments in the area
# however we make it part of 중구 for ease of use in manipulating data with consistency
# 여주군 has changed to 여주시
# 청원군 has become part of 청주시
# 연기군 has become part of 세종특별자치시
# 당진군 has changed to 당진시
# 마산시, 진해시 have become part of 창원시
# 제주도 has changed to 제주특별자치도

# creating resources to fix area codes
def area_code_fix(x):
    if x in area_code_fix_map.keys():
        return area_code_fix_map[x]
    else:
        return x

area_code_fix_map = {
    28720: 28110, # 용진군: 중구
    41730: 41670, # 여주군: 여주시
    43710: 43110, # 청원군: 청주시
    44730: 36110, # 연기군: 세종특별자치시
    44830: 44270, # 당진군: 당진시
    48150: 48110, # 마산시: 창원시
    48190: 48110, # 진해시: 창원시
    49110: 50110, # 제주시: 제주시
    49130: 50130, # 서귀포시: 서귀포시
    49710: 50110, # 북제주군: 제주시
    49720: 50130, # 남제주군: 서귀포시
}

df_pop.area_code = df_pop.area_code.apply(area_code_fix)

# one last sanity check
estate_area_code = df_estate.area_code[~df_estate.area_code.duplicated()]
estate_area_code_missing = estate_area_code[~estate_area_code.isin(df_pop.area_code)]
estate_area_code_missing.values

pop_area_code = df_pop.area_code[~df_pop.area_code.duplicated()]
pop_area_code_missing = pop_area_code[~pop_area_code.isin(df_estate.area_code)]
pop_area_code_missing.values

# exporting scrubbed dataframes
df_estate.to_csv(path + '03_estate.csv', index=False, encoding='utf-8')
df_pop.to_csv(path + '03_pop.csv', index=False, encoding='utf-8')
