==> 07_data_aggregation.ipynb <==
# 파트 07
----
앞부분의 전처리 된 파일들을 마지막으로 하나로 합쳐서 출력한다. 이 파트에서도 자잘한 추가적인 전처리가 이루어진다. 우선적으로 인구이동 데이터와 부동산 데이터 사이의 공통된 단위들로 묶는다. 종합 후의 파일 크기가 크지 않은 점과 추후 다른파트에서 분석의 용이성을 위해 아래의 두가지 방법으로 종합한다.

* 행정구역, 년
* 행정구역, 년, 월

종합하는 과정에서 우선적으로 전입사유 데이터가 카테고리인 점을 고려하여 새로 인코딩을 한다. 각 칼럼을 단순 총합을 할 경우, 대다수 칼럼이 전입세대의 수와 비례함으로 인한 칼럼간의 공선성의 문제가 있으므로 개별 칼럼마다 적절한 종합함수를 정의하여 사용한다.

%matplotlib inline
import re
from __future__ import division
from __future__ import print_function

# setting the folder path and read in data
path = 'raw_data\\'
df_pop = pd.read_csv(path + '06_pop.csv', encoding='utf-8')
df_estate = pd.read_csv(path + '06_estate.csv', encoding='utf-8')

# create new columns for price per area and log(price per area)
df_estate['ppa'] = df_estate.price / df_estate.area

df_estate['ppa_mean'] = np.nan
df_estate['ppa_median'] = np.nan
df_estate['trade_vol'] = 0

# custom aggregate functions the real estate data
def ppa_mean_agg(series):
    return df_estate.loc[series.index, 'ppa'].mean()

def ppa_median_agg(series):
    return df_estate.loc[series.index, 'ppa'].median()   

estate_agg_func_dic = {
    'built': 'mean',
    'ppa_mean': ppa_mean_agg,
    'ppa_median': ppa_median_agg,
    'trade_vol': 'count',
}

df_estate_agg_year = df_estate.groupby(['year', 'area_code']).agg(estate_agg_func_dic)
df_estate_agg_month = df_estate.groupby(['year', 'month', 'area_code']).agg(estate_agg_func_dic)

df_estate_agg_year['ppa_median_log'] = np.log(df_estate_agg_year.ppa_median)
df_estate_agg_month['ppa_median_log'] = np.log(df_estate_agg_month.ppa_median)

df_estate_agg_year['ppa_change_pre']  = np.nan
df_estate_agg_year['ppa_change_post']  = np.nan

for i in df_estate_agg_year.index.levels[0][:-1]:
    df_estate_agg_year.loc[i].ppa_change_pre =\
        df_estate_agg_year.loc[i + 1].ppa_median / df_estate_agg_year.loc[i].ppa_median
    df_estate_agg_year.loc[i + 1].ppa_change_post =\
        df_estate_agg_year.loc[i + 1].ppa_median / df_estate_agg_year.loc[i].ppa_median

# just another sanity check sanity check
df_pop.loc[df_pop.reason == -1].head()

# remove entries with missing values
df_pop = df_pop.loc[df_pop.reason != -1]

# turning categorical values into binary vectors
df_r = pd.get_dummies(df_pop.reason, prefix='r')
df_pop = pd.concat([df_pop, df_r], axis=1).drop('reason', axis=1)

df_pop.head()

# getting ready for aggregation
df_pop['mov_vol'] = np.nan
df_pop['mov_ppl_vol'] = np.nan
df_pop['pop_change_ratio'] = np.nan
df_pop['head_ratio'] = np.nan
df_pop['head_male_ratio'] = np.nan
df_pop['ppl_per_house'] = np.nan
df_pop['mov_male_ratio'] = np.nan
df_pop['multi_ratio'] = np.nan
df_pop['sign'] = (-1) ** (df_pop.move_in + 1)

# custom aggregate functions for the population data
def mov_vol_agg(series):
    return df_pop.loc[series.index, 'house_total'].sum()

def mov_ppl_vol_agg(series):
    return df_pop.loc[series.index, 'moved_ppl'].sum()

def pop_change_ratio_agg(series):
    return ((df_pop.loc[series.index, 'moved_ppl'] * df_pop.loc[series.index, 'sign']).sum() /
             df_pop.loc[series.index, 'moved_ppl'].sum())

def head_ratio_agg(series):
    return df_pop.loc[series.index, 'head'].sum() / df_pop.loc[series.index, 'house_total'].sum()

def head_male_ratio_agg(series):
    return df_pop.loc[series.index, 'head_male'].sum() / df_pop.loc[series.index, 'head'].sum()

def ppl_per_house_agg(series):
    return df_pop.loc[series.index, 'moved_ppl'].sum() / df_pop.loc[series.index, 'house_total'].sum()

def mov_male_ratio_agg(series):
    return df_pop.loc[series.index, 'moved_male'].sum() / df_pop.loc[series.index, 'moved_ppl'].sum()

def head_age_agg(series):
    return ((df_pop.loc[series.index, 'head_age'] * df_pop.loc[series.index, 'head']).sum() /
             df_pop.loc[series.index, 'head'].sum())

def reason_ratio_agg(series):
    return ((series * df_pop.loc[series.index, 'house_total'] * df_pop.loc[series.index, 'sign']).sum() /
             df_pop.loc[series.index, 'house_total'].sum())

def multi_ratio_agg(series):
    return df_pop.loc[series.index, 'multi'].sum() / df_pop.loc[series.index, 'house_total'].sum()

pop_agg_func_dic = {
    'mov_vol': mov_vol_agg,
    'mov_ppl_vol': mov_ppl_vol_agg,
    'pop_change_ratio': pop_change_ratio_agg,
    'head_ratio': head_ratio_agg,
    'head_male_ratio': head_male_ratio_agg,
    'ppl_per_house': ppl_per_house_agg,
    'mov_male_ratio': mov_male_ratio_agg,
    'multi_ratio': multi_ratio_agg,
    'head_age': head_age_agg,
    'r_1': reason_ratio_agg,
    'r_2': reason_ratio_agg,
    'r_3': reason_ratio_agg,
    'r_4': reason_ratio_agg,
    'r_5': reason_ratio_agg,
    'r_6': reason_ratio_agg,
    'r_9': reason_ratio_agg,
}

# group things accordingly and aggregate the data using custom aggregate functions
df_pop_agg_year = df_pop.groupby(['year', 'area_code']).agg(pop_agg_func_dic)
df_pop_agg_month = df_pop.groupby(['year', 'month', 'area_code']).agg(pop_agg_func_dic)

# combine the population data and the real estate data into one
df_year = pd.concat([df_pop_agg_year, df_estate_agg_year], axis=1)
df_month = pd.concat([df_pop_agg_month, df_estate_agg_month], axis=1)

df_year.reset_index(inplace=True)
df_month.reset_index(inplace=True)

# see if there are any null values
df_year.loc[df_year.loc[:, 'built'].isnull()]

# null built values above indicate there weren't any houses sold during the period
# since we cannot do an analysis on the housing prices over these data points
# we drop these data points
df_year = df_year.loc[~df_year.loc[:, 'built'].isnull()]
df_month = df_month.loc[~df_month.loc[:, 'built'].isnull()]

df_year.to_csv(path + '07_pop_estate_year.csv', index=False, encoding='utf-8')
df_month.to_csv(path + '07_pop_estate_month.csv', index=False, encoding='utf-8')
